company_directory/backend/companies/serializers.py
from rest_framework import serializers
from .models import Company, Location

class LocationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Location
        fields = '__all__'

class CompanySerializer(serializers.ModelSerializer):
    locations = LocationSerializer(many=True, read_only=True)

    class Meta:
        model = Company
        fields = '__all__'
