company_directory/backend/requirements.txt
Django>=3.2
djangorestframework
psycopg2-binary
pandas
django-cors-headers

